coreapi/requirements.txt
fastapi==0.95.1
langchain==0.0.149
openai==0.27.4
pinecone-client==2.2.1
pinecone_text==0.4.2
python-dotenv==1.0.0
uvicorn==0.21.1
vocode==0.1.109
tiktoken==0.3.3
google-api-python-client==2.86.0
google-cloud-texttospeech==2.14.1
mysql-connector-python==8.0.33
line-bot-sdk==2.4.2
typing-inspect==0.8.0
typing_extensions==4.5.0
alembic==1.10.4
PyJWT==2.7.0
pytest==7.3.1
httpx==0.24.0
mypy==1.3.0
fastapi-utils==0.2.1
